FE refinements for sector field
