+++ web/src/Definitions/actionDefinitions.tsx
@@ -2,7 +2,6 @@ export type ActionMigratedData = {
   type: string[];
   ghgsAffected: string;
   natImplementer: string[];
-  sectorsAffected: string[];
   estimatedInvestment: number;
   achievedReduction: number;
   expectedReduction: number;
