+++ web/src/Pages/Programmes/ProgrammeForm/programmeForm.tsx
@@ -7,7 +7,6 @@ import { useNavigate, useParams } from 'react-router-dom';
 import UploadFileGrid from '../../../Components/Upload/uploadFiles';
 import AttachEntity from '../../../Components/Popups/attach';
 import { useConnection } from '../../../Context/ConnectionContext/connectionContext';
-import { Sector } from '../../../Enums/sector.enum';
 import { SubSector, NatImplementor } from '../../../Enums/shared.enum';
 import { ProgrammeStatus } from '../../../Enums/programme.enum';
 import { Layers } from 'react-bootstrap-icons';
@@ -28,7 +27,6 @@ import { getProjectTableColumns } from '../../../Definitions/columns/projectColu
 // import { ActivityData } from '../../../Definitions/activityDefinitions';
 // import { SupportData } from '../../../Definitions/supportDefinitions';
 import UpdatesTimeline from '../../../Components/UpdateTimeline/updates';
-import { useUserContext } from '../../../Context/UserInformationContext/userInformationContext';
 
 const { Option } = Select;
 const { TextArea } = Input;
@@ -46,7 +44,6 @@ const ProgrammeForm: React.FC<FormLoadProps> = ({ method }) => {
   const navigate = useNavigate();
   const { get, post, put } = useConnection();
   const ability = useAbilityContext();
-  const { userInfoState } = useUserContext();
   const { entId } = useParams();
 
   // Form Validation Rules
@@ -105,18 +102,10 @@ const ProgrammeForm: React.FC<FormLoadProps> = ({ method }) => {
 
   // Initialization Logic
 
-  const availableSectors: string[] = [];
-  const userSectors = userInfoState?.userSectors?.split(',') ?? [];
   const yearsList: number[] = [];
 
   for (let year = 2013; year <= 2050; year++) {
     yearsList.push(year);
-  }
-
-  if (userInfoState?.userRole === 'Root' || userInfoState?.userRole === 'Admin') {
-    Object.values(Sector).map((sector) => availableSectors.push(sector));
-  } else {
-    userSectors.map((sector) => availableSectors.push(sector));
   }
 
   useEffect(() => {
@@ -174,7 +163,7 @@ const ProgrammeForm: React.FC<FormLoadProps> = ({ method }) => {
               programmeStatus: entityData.programmeStatus,
               startYear: entityData.startYear,
               natAnchor: entityData.natAnchor,
-              affectedSectors: entityData.affectedSectors,
+              sector: entityData.sector,
               affectedSubSector: entityData.affectedSubSector,
               natImplementor: entityData.nationalImplementor,
               investment: entityData.investment,
@@ -722,23 +711,13 @@ const ProgrammeForm: React.FC<FormLoadProps> = ({ method }) => {
                 <Col span={6}>
                   <Form.Item
                     label={<label className="form-item-header">{t('sectorsAffTitle')}</label>}
-                    name="affectedSectors"
-                    rules={[validation.required]}
+                    name="sector"
                   >
                     <Select
                       size="large"
                       style={{ fontSize: inputFontSize }}
-                      mode="multiple"
-                      allowClear
-                      disabled={isView}
-                      showSearch
-                    >
-                      {availableSectors.map((sector) => (
-                        <Option key={sector} value={sector}>
-                          {sector}
-                        </Option>
-                      ))}
-                    </Select>
+                      disabled={true}
+                    ></Select>
                   </Form.Item>
                 </Col>
                 <Col span={6}>
